Use new activity names
