+++ packages/firebase_performance/example/android/app/src/main/java/io/flutter/plugins/firebaseperformanceexample/MainActivity.java
@@ -1,13 +1,12 @@
 package io.flutter.plugins.firebaseperformanceexample;
 
-import android.os.Bundle;
-import io.flutter.app.FlutterActivity;
-import io.flutter.plugins.GeneratedPluginRegistrant;
+import io.flutter.embedding.android.FlutterActivity;
+import io.flutter.embedding.engine.FlutterEngine;
+import io.flutter.plugins.firebaseperformance.FirebasePerformancePlugin;
 
 public class MainActivity extends FlutterActivity {
   @Override
-  protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    GeneratedPluginRegistrant.registerWith(this);
+  public void configureFlutterEngine(FlutterEngine flutterEngine) {
+    flutterEngine.getPlugins().add(new FirebasePerformancePlugin());
   }
 }
