--- packages/firebase_performance/example/android/app/src/main/java/io/flutter/plugins/firebaseperformanceexample/MainActivity2.java
@@ -1,12 +1,0 @@
-package io.flutter.plugins.firebaseperformanceexample;
-
-import io.flutter.embedding.android.FlutterActivity;
-import io.flutter.embedding.engine.FlutterEngine;
-import io.flutter.plugins.firebaseperformance.FirebasePerformancePlugin;
-
-public class MainActivity2 extends FlutterActivity {
-  @Override
-  public void configureFlutterEngine(FlutterEngine flutterEngine) {
-    flutterEngine.getPlugins().add(new FirebasePerformancePlugin());
-  }
-}
